Move pinchAction() to utils
